[tr wpk-training/2791] Media - Pre Embauche - PFE 2023 - UBO - Réalisation - Back - Modules - WpRole - Autorisations

+++ app/Http/Controllers/WpRoleController.php
@@ -6,8 +6,11 @@ use App\Http\Requests\StoreWpRoleRequest;
 use App\Http\Requests\UpdateWpRoleRequest;
 use App\Models\WpRole;
 use App\Services\WpRoleService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class WpRoleController extends Controller
 {
@@ -39,6 +42,9 @@ class WpRoleController extends Controller
      */
     public function store(StoreWpRoleRequest $request): JsonResponse
     {
+        if ($request->user()->cannot('createWpRole', WpRole::class)) {
+            abort(403);
+        }
         return $this->wpRoleService->storeWpRole($request->all());
     }
     
@@ -61,6 +67,9 @@ class WpRoleController extends Controller
      */
     public function update(UpdateWpRoleRequest $request, WpRole $wpRole): JsonResponse
     {
+        if ($request->user()->cannot('updateWpRole', WpRole::class)) {
+            abort(403);
+        }
         return $this->wpRoleService->updateWpRole($request->all(), $wpRole);
     }
         
@@ -72,6 +81,9 @@ class WpRoleController extends Controller
      */
     public function destroy(WpRole $wpRole): JsonResponse
     {
+        if(Auth::user()->cannot('delete', WpRole::class)){
+            abort(403); 
+        }
         return $this->wpRoleService->deleteWpRole($wpRole);
     }
 }
